Add prompt for simulator install, and clean up output of xcode checks.
