+++ src/briefcase/platforms/iOS/xcode.py
@@ -1,5 +1,6 @@
 import subprocess
 import time
+from pathlib import Path
 from uuid import UUID
 
 from briefcase.commands import (
